% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keen.R
\name{ch_navbar}
\alias{ch_navbar}
\title{Create a navbar at the top of the page}
\usage{
ch_navbar(brand, ...)
}
\arguments{
\item{brand}{title to display on far left of navbar}

\item{...}{other elements in navbar}
}
\description{
Create a navbar at the top of the page
}
