% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keen.R
\name{ch_panel}
\alias{ch_panel}
\title{Create a chart panel}
\usage{
ch_panel(chart, title = "", notes = "", id = paste0("panel-",
  round(runif(1) * 10000, 0)))
}
\arguments{
\item{chart}{chart object to wrap}

\item{title}{panel title to be displayed at the top}

\item{notes}{panel notes to be displayed at the bottom}

\item{id}{id for panel element}

\item{cols}{grid width of the column. can be a named vector to allow for
different column sizes based on screen size.}
}
\description{
Create a chart panel
}
