% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keen.R
\name{Navbar2}
\alias{Navbar2}
\title{Create a navbar at the top of the page}
\usage{
Navbar2(brand, ...)
}
\arguments{
\item{brand}{title to display on far left of navbar}

\item{...}{other elements in navbar}
}
\description{
Create a navbar at the top of the page
}
