% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{panel_title}
\alias{panel_title}
\title{Create a panel title}
\usage{
panel_title(title, note = "", idright)
}
\arguments{
\item{title}{title}

\item{note}{note}

\item{idright}{id of span element created to the far right}
}
\description{
Create a panel title
}
